+++ src/web3id/v1.rs
@@ -1,0 +1,2 @@
+/// Functionality to create the verification request anchor (VRA) and verification audit anchor (VAA).
+pub mod anchor;